drivers: counter: Fix possible out-of-bounds access in MCP7940N driver
This PR fixes the issue of possible out-of-bounds
access if write_data_block function was misused.

Size argument of the function was only validated
against the maximum allowed value, not the real
size of the data struct to be read, what could
cause accessing data past the boundary of the struct.

Should fix #81930.

Signed-off-by: Marcin Lyda <elektromarcin@gmail.com>